requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
python-dotenv
dbt-core
dbt-redshift
astronomer-cosmos
apache-airflow-providers-amazon